update metadata structure
