+++ src/ibl_mesoscope_to_nwb/mesoscope2025/metadata/processed_ophys_metadata.yaml
